+++ src/Player/SplashManager.cpp
@@ -104,7 +104,7 @@ void TSplashManager::makeDL(JDrama::TGraphics* param_1) const
 		f32 fVar1 = ((f32)unk648 - splash->unk10) / unk648 * unk634 + unk630;
 
 		GXColor color
-		    = (GXColor) { 0xff, 0xff, 0xff, splash->unk10 * 255 / unk648 };
+		    = (GXColor) { 0xff, 0xff, 0xff, (u8)(splash->unk10 * 255 / unk648) };
 
 		thing[0].set(pos.x - fVar1, pos.y + fVar1, pos.z);
 		thing[1].set(pos.x + fVar1, pos.y + fVar1, pos.z);
